refactor(capabilities): split into separate files with collocated tools
Reorganize capabilities module from a single file into a directory structure
where each capability has its own file with its tools collocated. This improves
code organization and makes it easier to work with individual capabilities.

Directory structure:
- capabilities/mod.rs: Capability trait, registry, apply_capabilities, re-exports
- capabilities/noop.rs: NoopCapability
- capabilities/current_time.rs: CurrentTimeCapability + GetCurrentTimeTool
- capabilities/research.rs: ResearchCapability (ComingSoon)
- capabilities/sandbox.rs: SandboxCapability (ComingSoon)
- capabilities/file_system.rs: FileSystemCapability (ComingSoon)
- capabilities/test_math.rs: TestMathCapability + 4 math tools
- capabilities/test_weather.rs: TestWeatherCapability + 2 weather tools
- capabilities/stateless_todo_list.rs: StatelessTodoListCapability + WriteTodosTool
