+++ crates/everruns-core/src/capabilities/current_time.rs
@@ -1,0 +1,111 @@
+//! CurrentTime Capability - provides tools to get current date and time
+
+use super::{Capability, CapabilityId, CapabilityStatus};
+use crate::tools::{Tool, ToolExecutionResult};
+use async_trait::async_trait;
+use serde_json::Value;
+
+/// CurrentTime capability - provides tools to get current date and time
+pub struct CurrentTimeCapability;
+
+impl Capability for CurrentTimeCapability {
+    fn id(&self) -> &str {
+        CapabilityId::CURRENT_TIME
+    }
+
+    fn name(&self) -> &str {
+        "Current Time"
+    }
+
+    fn description(&self) -> &str {
+        "Adds a tool to get the current date and time in various formats and timezones."
+    }
+
+    fn status(&self) -> CapabilityStatus {
+        CapabilityStatus::Available
+    }
+
+    fn icon(&self) -> Option<&str> {
+        Some("clock")
+    }
+
+    fn category(&self) -> Option<&str> {
+        Some("Utilities")
+    }
+
+    fn tools(&self) -> Vec<Box<dyn Tool>> {
+        vec![Box::new(GetCurrentTimeTool)]
+    }
+}
+
+// ============================================================================
+// Tool: get_current_time
+// ============================================================================
+
+/// Tool that returns the current date and time
+pub struct GetCurrentTimeTool;
+
+#[async_trait]
+impl Tool for GetCurrentTimeTool {
+    fn name(&self) -> &str {
+        "get_current_time"
+    }
+
+    fn description(&self) -> &str {
+        "Get the current date and time. Can return time in different formats and timezones."
+    }
+
+    fn parameters_schema(&self) -> Value {
+        serde_json::json!({
+            "type": "object",
+            "properties": {
+                "timezone": {
+                    "type": "string",
+                    "description": "Timezone to return the time in (e.g., 'UTC', 'America/New_York', 'Europe/London'). Defaults to UTC."
+                },
+                "format": {
+                    "type": "string",
+                    "enum": ["iso8601", "unix", "human"],
+                    "description": "Output format: 'iso8601' for ISO 8601 format, 'unix' for Unix timestamp, 'human' for human-readable format. Defaults to 'iso8601'."
+                }
+            },
+            "additionalProperties": false
+        })
+    }
+
+    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
+        let format = arguments
+            .get("format")
+            .and_then(|v| v.as_str())
+            .unwrap_or("iso8601");
+
+        let _timezone = arguments
+            .get("timezone")
+            .and_then(|v| v.as_str())
+            .unwrap_or("UTC");
+
+        // Note: For simplicity, we're using UTC. Full timezone support would require
+        // the chrono-tz crate which adds significant dependencies.
+        let now = chrono::Utc::now();
+
+        let result = match format {
+            "unix" => serde_json::json!({
+                "timestamp": now.timestamp(),
+                "format": "unix",
+                "timezone": "UTC"
+            }),
+            "human" => serde_json::json!({
+                "datetime": now.format("%A, %B %d, %Y at %H:%M:%S UTC").to_string(),
+                "format": "human",
+                "timezone": "UTC"
+            }),
+            _ => serde_json::json!({
+                "datetime": now.to_rfc3339(),
+                "format": "iso8601",
+                "timezone": "UTC"
+            }),
+        };
+
+        ToolExecutionResult::success(result)
+    }
+}
