+++ crates/everruns-core/src/capabilities/test_math.rs
@@ -1,0 +1,249 @@
+//! TestMath Capability - calculator tools for testing tool calling
+
+use super::{Capability, CapabilityId, CapabilityStatus};
+use crate::tools::{Tool, ToolExecutionResult};
+use async_trait::async_trait;
+use serde_json::Value;
+
+/// TestMath capability - calculator tools for testing tool calling
+pub struct TestMathCapability;
+
+impl Capability for TestMathCapability {
+    fn id(&self) -> &str {
+        CapabilityId::TEST_MATH
+    }
+
+    fn name(&self) -> &str {
+        "Test Math"
+    }
+
+    fn description(&self) -> &str {
+        "Testing capability: adds calculator tools (add, subtract, multiply, divide) for tool calling tests."
+    }
+
+    fn status(&self) -> CapabilityStatus {
+        CapabilityStatus::Available
+    }
+
+    fn icon(&self) -> Option<&str> {
+        Some("calculator")
+    }
+
+    fn category(&self) -> Option<&str> {
+        Some("Testing")
+    }
+
+    fn system_prompt_addition(&self) -> Option<&str> {
+        Some("You have access to math tools. Use them for calculations: add, subtract, multiply, divide.")
+    }
+
+    fn tools(&self) -> Vec<Box<dyn Tool>> {
+        vec![
+            Box::new(AddTool),
+            Box::new(SubtractTool),
+            Box::new(MultiplyTool),
+            Box::new(DivideTool),
+        ]
+    }
+}
+
+// ============================================================================
+// Tool: add
+// ============================================================================
+
+/// Tool that adds two numbers
+pub struct AddTool;
+
+#[async_trait]
+impl Tool for AddTool {
+    fn name(&self) -> &str {
+        "add"
+    }
+
+    fn description(&self) -> &str {
+        "Add two numbers together and return the result."
+    }
+
+    fn parameters_schema(&self) -> Value {
+        serde_json::json!({
+            "type": "object",
+            "properties": {
+                "a": {
+                    "type": "number",
+                    "description": "The first number"
+                },
+                "b": {
+                    "type": "number",
+                    "description": "The second number"
+                }
+            },
+            "required": ["a", "b"],
+            "additionalProperties": false
+        })
+    }
+
+    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
+        let a = arguments.get("a").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let b = arguments.get("b").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let result = a + b;
+
+        ToolExecutionResult::success(serde_json::json!({
+            "result": result,
+            "operation": "add",
+            "a": a,
+            "b": b
+        }))
+    }
+}
+
+// ============================================================================
+// Tool: subtract
+// ============================================================================
+
+/// Tool that subtracts two numbers
+pub struct SubtractTool;
+
+#[async_trait]
+impl Tool for SubtractTool {
+    fn name(&self) -> &str {
+        "subtract"
+    }
+
+    fn description(&self) -> &str {
+        "Subtract the second number from the first and return the result."
+    }
+
+    fn parameters_schema(&self) -> Value {
+        serde_json::json!({
+            "type": "object",
+            "properties": {
+                "a": {
+                    "type": "number",
+                    "description": "The number to subtract from"
+                },
+                "b": {
+                    "type": "number",
+                    "description": "The number to subtract"
+                }
+            },
+            "required": ["a", "b"],
+            "additionalProperties": false
+        })
+    }
+
+    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
+        let a = arguments.get("a").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let b = arguments.get("b").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let result = a - b;
+
+        ToolExecutionResult::success(serde_json::json!({
+            "result": result,
+            "operation": "subtract",
+            "a": a,
+            "b": b
+        }))
+    }
+}
+
+// ============================================================================
+// Tool: multiply
+// ============================================================================
+
+/// Tool that multiplies two numbers
+pub struct MultiplyTool;
+
+#[async_trait]
+impl Tool for MultiplyTool {
+    fn name(&self) -> &str {
+        "multiply"
+    }
+
+    fn description(&self) -> &str {
+        "Multiply two numbers together and return the result."
+    }
+
+    fn parameters_schema(&self) -> Value {
+        serde_json::json!({
+            "type": "object",
+            "properties": {
+                "a": {
+                    "type": "number",
+                    "description": "The first number"
+                },
+                "b": {
+                    "type": "number",
+                    "description": "The second number"
+                }
+            },
+            "required": ["a", "b"],
+            "additionalProperties": false
+        })
+    }
+
+    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
+        let a = arguments.get("a").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let b = arguments.get("b").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let result = a * b;
+
+        ToolExecutionResult::success(serde_json::json!({
+            "result": result,
+            "operation": "multiply",
+            "a": a,
+            "b": b
+        }))
+    }
+}
+
+// ============================================================================
+// Tool: divide
+// ============================================================================
+
+/// Tool that divides two numbers
+pub struct DivideTool;
+
+#[async_trait]
+impl Tool for DivideTool {
+    fn name(&self) -> &str {
+        "divide"
+    }
+
+    fn description(&self) -> &str {
+        "Divide the first number by the second and return the result. Returns an error if dividing by zero."
+    }
+
+    fn parameters_schema(&self) -> Value {
+        serde_json::json!({
+            "type": "object",
+            "properties": {
+                "a": {
+                    "type": "number",
+                    "description": "The dividend (number to be divided)"
+                },
+                "b": {
+                    "type": "number",
+                    "description": "The divisor (number to divide by)"
+                }
+            },
+            "required": ["a", "b"],
+            "additionalProperties": false
+        })
+    }
+
+    async fn execute(&self, arguments: Value) -> ToolExecutionResult {
+        let a = arguments.get("a").and_then(|v| v.as_f64()).unwrap_or(0.0);
+        let b = arguments.get("b").and_then(|v| v.as_f64()).unwrap_or(0.0);
+
+        if b == 0.0 {
+            return ToolExecutionResult::tool_error("Cannot divide by zero");
+        }
+
+        let result = a / b;
+
+        ToolExecutionResult::success(serde_json::json!({
+            "result": result,
+            "operation": "divide",
+            "a": a,
+            "b": b
+        }))
+    }
+}
